Refactoring to prevent interactions with other ActiveRecord additions.
  * Drop old plugin style code
  * Drop Rails 3.2.x compatibility
  * Use railtie to cleanly apply additions to ActiveRecord::Base
  * Rename translations methods to has_translation to prevent shadowing with translations relation.
