+++ .travis.yml
@@ -3,13 +3,11 @@ rvm:
   - 1.9.2
   - 1.9.3
 env:
-  - "RAILS_VERSION=2.3.0"
   - "RAILS_VERSION=3.0.0"
   - "RAILS_VERSION=3.1.0"
   - "RAILS_VERSION=3.2.0"
 
 gemfile:
-  - gemfiles/Gemfile.rails-2.3
   - gemfiles/Gemfile.rails-3.0
   - gemfiles/Gemfile.rails-3.1
   - gemfiles/Gemfile.rails-3.2
