--- gemfiles/Gemfile.rails-2.3
@@ -1,1 +1,0 @@
-gem 'rails', '~> 2.3'
