--- init.rb
@@ -1,1 +1,0 @@
-require 'has_translations'
