+++ lib/has_translations/version.rb
@@ -1,3 +1,3 @@
 module HasTranslations
-  VERSION = "0.3.5"
+  VERSION = "1.0.0.alpha.1"
 end
